+++ lib/src/verifyJWT/ivm_check_integrity.dart
@@ -35,14 +35,16 @@ Future<Map<String, dynamic>> _checkTokenIntegrity(String token) async {
   /// Base64url-decode the Header, ensuring that no line breaks, whitespace, or other additional characters have been used, and verify that the decoded Header is a valid JSON object.
   ///
   try {
-    jwtHeader = await Utilities.base64UrlDecode(tokenSegments[0]);
+    jwtHeader =
+        utf8.decode(base64Url.decode(base64Url.normalize(tokenSegments[0])));
   } catch (e) {
     throw Exception('Error decoding header segment! $e.');
   }
 
   // Verify if the header is a valid JSON
   try {
-    validHeader = JsonValidator(jwtHeader).validate();
+    final jv = JsonValidator();
+    validHeader = await jv.validate(jwtHeader);
     // validHeader = await Utilities.validateSegmentToJSON(jwtHeader);
   } catch (e) {
     throw Exception('Error validating header segment! $e.');
@@ -51,15 +53,16 @@ Future<Map<String, dynamic>> _checkTokenIntegrity(String token) async {
   /// Base64url-decode the Payload, ensuring that no line breaks, whitespace, or other additional characters have been used, and verify that the decoded Payload is a valid JSON object.
   ///
   try {
-    jwtPayload = await Utilities.base64UrlDecode(tokenSegments[1]);
+    jwtPayload =
+        utf8.decode(base64Url.decode(base64Url.normalize(tokenSegments[1])));
   } catch (e) {
     throw Exception('Error decoding payload segment! $e.');
   }
 
   // Verify if the payload is a valid JSON
   try {
-    validPayload = JsonValidator(jwtPayload).validate();
-    // validPayload = await Utilities.validateSegmentToJSON(jwtPayload);
+    final jv = JsonValidator();
+    validPayload = await jv.validate(jwtPayload);
   } catch (e) {
     throw Exception('Error validating payload segment! $e.');
   }