+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+v. 0.1.4    fixes in base64Url encode decode
+
+v. 0.1.3    extended tests. syntax fixes.
+
+v. 0.1.2    fixes for async jvalid.
+
 v. 0.1.1    few minor fixes.
+
 v. 0.1.0    first beta release.
             Issue RSA [SHA-256] signed JWTokens. Verify JWToken signed with RSA [SHA-256].
+
 v. 0.0.1    initial work started.